requirements.txt
flask==3.0.3
pandas==2.2.2
pygtrie==2.5.0
python-dotenv==1.0.1
